requirements.txt
absl-py
matplotlib
numpy
scipy
pillow